restructure package for easier installation and imports, fix and update imports and pendulum dependency
