+++ requirements.txt
@@ -4,5 +4,5 @@ matplotlib
 multiprocess
 numpy
 pandas
-pendulum
+pendulum>=3.0.0
 scipy
