+++ src/superleaf/operators/__init__.py
@@ -1,4 +1,4 @@
-from superleaf.operators.base import BooleanFunctionOperator, FunctionOperator, Operator, operator, bool_operator
+from superleaf.operators.base import bool_operator, operator, Operator
 from superleaf.operators.comparison import ComparisonFunctions
 from superleaf.operators.getters import attr_getter, index_getter
 from superleaf.operators.string import str_bool_op, str_op
