+++ src/superleaf/operators/base.py
