+++ src/superleaf/plotting/intervals.py
