+++ src/superleaf/timeseries/time_intervals.py
